// ui-angular/src/lib/components/bmb-modal/bmb-modal.interface.ts
import { TemplateRef, Type } from '@angular/core';
import { IButtonAppearance } from '../../_shared/types';

export type IBmbModalSize = 'small' | 'medium' | 'large';
export type IBmbModalType = 'informative' | 'action' | 'alert';
export type IBmbNativeModalSize = IBmbModalSize | 'x-small' | 'x-large';
export type IBmbModalAlertStyle =
  | 'warning'
  | 'neutral'
  | 'primary'
  | 'event'
  | 'success'
  | 'error';
export interface IBmbActionButton {
  buttonName: string;
  appearance?: IButtonAppearance;
  label: string;
  icon?: string;
  action: (event?: Event, btn?: IBmbActionButton) => void;
}

export interface ModalDataConfig {
  title?: string;
  subtitle?: string;
  content?: string | TemplateRef<any> | null;
  type?: IBmbModalType;
  alertStyle?: IBmbModalAlertStyle;
  size?: IBmbModalSize;
  primaryBtnLabel?: string;
  secondaryBtnLabel?: string;
  scrollable?: boolean;
  hidePrimaryButton?: boolean;
  hideSecondaryButton?: boolean;
  extendButtons?: boolean;
  primaryAction?: () => void;
  secondaryAction?: () => void;
  closeAction?: () => void;
}
export interface IBmbNativeModal
  extends Omit<
    ModalDataConfig,
    | 'primaryAction'
    | 'secondaryAction'
    | 'closeAction'
    | 'extendButtons'
    | 'alertStyle'
    | 'size'
    | 'type'
    | 'content'
  > {
  modalId?: string;
  size?: IBmbNativeModalSize;
  iconStyle?: IBmbModalAlertStyle;
  actions?: IBmbActionButton[];
  closeModalClicked?: (event: unknown) => void;
  beforeCloseModal?: (event: {
    modalId: string;
    reason: 'single' | 'all';
  }) => void;
  afterCloseModal?: (event: {
    modalId: string;
    reason: 'single' | 'all';
  }) => void;
  afterOpenModal?: (event: {
    modalId: string;
    reason: 'single' | 'all';
  }) => void;
  inputContext?: { [key: string]: any };
  outputContext?: { [key: string]: (value: any) => void };
  content?: string | TemplateRef<any> | null | Type<any>;
}